add get,put,monitor no block; lots of cleanup; more needed
