+++ documentation/exampleCPP.html
@@ -26,7 +26,7 @@
 
 <div class="head">
 <h1>EPICS exampleCPP</h1>
-<h2 class="nocount">Release 4.3.0-SNAPSHOT - 2016.09.02</h2>
+<h2 class="nocount">Release 4.4.0-SNAPSHOT - 2017.07.12</h2>
 
 <h2 class="nocount">Abstract</h2>
 <p>
