+++ exampleClient/src/examplePvaClientMonitor.cpp
@@ -9,8 +9,6 @@
 
 /* Author: Marty Kraimer */
 
-#include <epicsThread.h>
-
 #include <iostream>
 
 #include <pv/pvaClient.h>
@@ -54,7 +52,6 @@ int main(int argc,char *argv[])
     cout << "_____examplePvaClientMonitor starting_______\n";
     try {
         PvaClientPtr pva = PvaClient::get("pva ca");
-//PvaClient::setDebug(true);
         exampleMonitor(pva,"PVRdouble","pva");
         PvaClientChannelPtr pvaChannel = pva->createChannel("DBRdouble00","ca");
         pvaChannel->issueConnect();
