+++ exampleClient/src/putNoBlock.cpp
@@ -11,25 +11,27 @@
 #include <iostream>
 
 #include <pv/pvaClient.h>
+#include <pv/convert.h>
 
 using namespace std;
 using namespace epics::pvData;
 using namespace epics::pvAccess;
 using namespace epics::pvaClient;
 
-class ClientGet;
-typedef std::tr1::shared_ptr<ClientGet> ClientGetPtr;
+class ClientPut;
+typedef std::tr1::shared_ptr<ClientPut> ClientPutPtr;
 
-class ClientGet :
+class ClientPut :
     public PvaClientChannelStateChangeRequester
 {
 private:
     string request;
     PvaClientChannelPtr pvaClientChannel;
-    PvaClientGetPtr pvaClientGet;
+    PvaClientPutPtr pvaClientPut;
+    PvaClientPutDataPtr putData;
 public:
-    POINTER_DEFINITIONS(ClientGet);
-    ClientGet(
+    POINTER_DEFINITIONS(ClientPut);
+    ClientPut(
         const string &request,
         const PvaClientChannelPtr & pvaClientChannel)
     : request(request),
@@ -37,32 +39,38 @@ public:
     {
     }
     
-    static ClientGetPtr create(
+    static ClientPutPtr create(
         PvaClientPtr const &pvaClient,
         const string & channelName,
         const string & providerName,
         const string  & request)
     {
         PvaClientChannelPtr pvaClientChannel = pvaClient->createChannel(channelName,providerName);
-        ClientGetPtr clientGet(new ClientGet(request,pvaClientChannel));
-        pvaClientChannel->setStateChangeRequester(clientGet);
+        ClientPutPtr clientPut(new ClientPut(request,pvaClientChannel));
+        pvaClientChannel->setStateChangeRequester(clientPut);
         pvaClientChannel->issueConnect();
-        return clientGet;
+        return clientPut;
     }
 
     virtual void channelStateChange(PvaClientChannelPtr const & channel, bool isConnected)
     {
-        if(isConnected&&!pvaClientGet)
+        if(isConnected&&!pvaClientPut)
         {
-           pvaClientGet  = pvaClientChannel->createGet(request);
-           pvaClientGet->issueConnect();
+           pvaClientPut  = pvaClientChannel->createPut(request);
+           pvaClientPut->issueConnect();
         }
     }
 
-    PvaClientGetPtr getPvaClientGet()
+    PvaClientPutPtr getPvaClientPut()
     {
-        return  pvaClientGet;
-    }   
+        return  pvaClientPut;
+    }
+
+    PvaClientPutDataPtr getData()
+    {
+         if(!putData) putData = pvaClientPut->getData();
+         return putData;
+    }
 };
 
 
@@ -70,7 +78,7 @@ int main(int argc,char *argv[])
 {
     string providerName("pva");
     string channelName("PVRdouble");
-    string request("value,alarm,timeStamp");
+    string request("value");
     bool debug(false);
     if(argc==2 && string(argv[1])==string("-help")) {
         cout << "providerName channelName request debug" << endl;
@@ -100,12 +108,13 @@ int main(int argc,char *argv[])
          << " request " << request
          << " debug " << (debug ? "true" : "false") << endl;
 
-    cout << "_____getForever starting__\n";
+    cout << "_____putNoBlock starting__\n";
     
     try {   
         if(debug) PvaClient::setDebug(true);
+        ConvertPtr convert = getConvert();
         PvaClientPtr pva= PvaClient::get(providerName);
-        ClientGetPtr clientGet(ClientGet::create(pva,channelName,providerName,request));
+        ClientPutPtr clientPut(ClientPut::create(pva,channelName,providerName,request));
         while(true) {
             int c = std::cin.peek();  // peek character
             if ( c == EOF ) continue;
@@ -113,20 +122,19 @@ int main(int argc,char *argv[])
             string str;
             getline(cin,str);
             if(str.compare("exit")==0) break;
-            PvaClientGetPtr pvaClientGet = clientGet->getPvaClientGet();
-            if(!pvaClientGet) {
+            PvaClientPutPtr pvaClientPut = clientPut->getPvaClientPut();
+            if(!pvaClientPut) {
                 cout << "not connected\n";
             } else {
-                pvaClientGet->get();
-                PvaClientGetDataPtr data = pvaClientGet->getData();
-                BitSetPtr bitSet =  data->getChangedBitSet();
-                if(bitSet->cardinality()>0) {
-                    cout << "changed\n";
-                    data->showChanged(cout);
-                    cout << "bitSet " << *bitSet << endl;
+                try {
+                    PvaClientPutDataPtr putData = clientPut->getData();
+                    PVScalarPtr pvScalar(putData->getScalarValue());
+                    convert->fromString(pvScalar,str);
+                    pvaClientPut->put();
+                } catch (std::runtime_error e) {
+                     cout << "str " << str << " is not a valid value\n";
                 }
             }
-            
         }
     } catch (std::runtime_error e) {
         cerr << "exception " << e.what() << endl;
