+++ exampleClient/src/putUnion.cpp
