--- exampleClient/src/testClientFactory.cpp
@@ -1,48 +1,0 @@
-/*
- * Copyright information and license terms for this software can be
- * found in the file LICENSE that is included with the distribution
- */
-
-/**
- * @author mrk
- */
-
-/* Author: Marty Kraimer */
-
-#include <epicsThread.h>
-
-#include <iostream>
-
-#include <pv/pvaClient.h>
-#include <pv/clientFactory.h>
-
-using namespace std;
-using namespace epics::pvData;
-using namespace epics::pvAccess;
-using namespace epics::pvaClient;
-
-
-
-
-int main(int argc,char *argv[])
-{
-
-    cout << "_____testClientFactory starting_______\n";
-    
-    try {
-        PvaClient::setDebug(true);
-        PvaClientPtr pva = PvaClient::get("pva");
-        while(true) {
-            cout << "Type exit to stop: \n";
-            string str;
-            getline(cin,str);
-            if(str.compare("exit")==0){
-                 break;
-            }
-        }
-    } catch (std::runtime_error e) {
-            cerr << "exception " << e.what() << endl;
-            return 1;
-    }
-    return 0;
-}
