+++ exampleLink/iocBoot/exampleLink/st.extca
@@ -11,7 +11,6 @@ dbLoadRecords("db/ai.db","name=exampleLinkAI");
 
 cd ${TOP}/iocBoot/${IOC}
 iocInit()
-startPVAServer local
 exampleMonitorLinkCreateRecord ca exampleMonitorLink doubleArrayExternal 
 exampleGetLinkCreateRecord ca exampleGetLink doubleArrayExternal
 examplePutLinkCreateRecord ca examplePutLink doubleArrayExternal
